requirements.txt
django
django-allauth
django-crispy-forms
python-decouple
# psycopg2